Refactor: move calc_purge_upto() to LogHandler
